+++ SuperNewRoles/Roles/Attribute/Debugger.cs
@@ -255,7 +255,7 @@ public static class Debugger
 
             ShapeshifterPanel panel = GameObject.Instantiate(minigame.PanelPrefab, minigame.transform);
             panel.transform.localScale *= 0.5f;
-            panel.transform.localPosition = new Vector3(minigame.XStart + (float)num * 1.45f, minigame.YStart + 0.8f, -1f);
+            panel.transform.localPosition = new Vector3(minigame.XStart + num * 1.45f, minigame.YStart + 0.8f, -1f);
             static void Create(ShapeshifterPanel panel, int index, Action action)
             {
                 panel.SetPlayer(index, CachedPlayer.LocalPlayer.Data, (Action)(() =>
@@ -287,7 +287,7 @@ public static class Debugger
             int num = index % 3;
             int num2 = index / 3;
             ShapeshifterPanel panel = GameObject.Instantiate(minigame.PanelPrefab, minigame.transform);
-            panel.transform.localPosition = new Vector3(minigame.XStart + (float)num * minigame.XOffset, minigame.YStart + (float)num2 * minigame.YOffset, -1f);
+            panel.transform.localPosition = new Vector3(minigame.XStart + num * minigame.XOffset, minigame.YStart + num2 * minigame.YOffset, -1f);
             static void Create(ShapeshifterPanel panel, int index, Action action)
             {
                 panel.SetPlayer(index, CachedPlayer.LocalPlayer.Data, (Action)(() =>
@@ -320,7 +320,7 @@ public static class Debugger
             int num2 = index / 3;
 
             ShapeshifterPanel panel = GameObject.Instantiate(minigame.PanelPrefab, minigame.transform);
-            panel.transform.localPosition = new Vector3(minigame.XStart + (float)num * minigame.XOffset, minigame.YStart + (float)num2 * minigame.YOffset, -1f);
+            panel.transform.localPosition = new Vector3(minigame.XStart + num * minigame.XOffset, minigame.YStart + num2 * minigame.YOffset, -1f);
             panel.SetPlayer(index, p.Data, (Action)(() =>
             {
                 if (MeetingHud.Instance != null) MeetingHud.Instance.transform.FindChild("ButtonStuff").gameObject.SetActive(true);
@@ -347,7 +347,7 @@ public static class Debugger
 
             ShapeshifterPanel panel = GameObject.Instantiate(minigame.PanelPrefab, minigame.transform);
             panel.transform.localScale *= 0.4f;
-            panel.transform.localPosition = new Vector3(minigame.XStart + (float)num * 1.5f, minigame.YStart + 0.45f, -1f);
+            panel.transform.localPosition = new Vector3(minigame.XStart + num * 1.5f, minigame.YStart + 0.45f, -1f);
             static void Create(ShapeshifterPanel panel, int index, Action action)
             {
                 panel.SetPlayer(index, CachedPlayer.LocalPlayer.Data, (Action)(() =>
@@ -385,7 +385,7 @@ public static class Debugger
 
             ShapeshifterPanel panel = GameObject.Instantiate(minigame.PanelPrefab, minigame.transform);
             panel.transform.localScale *= 0.3f;
-            panel.transform.localPosition = new Vector3(minigame.XStart - 0.7f + (float)num * 1f, minigame.YStart + (float)num2 * -0.3f, -1f);
+            panel.transform.localPosition = new Vector3(minigame.XStart - 0.7f + num * 1f, minigame.YStart + num2 * -0.3f, -1f);
             panel.SetPlayer(index, CachedPlayer.LocalPlayer.Data, (Action)(() =>
             {
                 if (MeetingHud.Instance != null) MeetingHud.Instance.transform.FindChild("ButtonStuff").gameObject.SetActive(true);
